Board and Drivers update from HTC 2.17 source release

+++ arch/arm/mach-tegra/pm.h
@@ -61,7 +61,7 @@ struct tegra_suspend_platform_data {
 	/* lp_state = 0 for LP0 state, 1 for LP1 state, 2 for LP2 state */
 	void (*board_resume)(int lp_state, enum resume_stage stg);
 	unsigned int cpu_resume_boost;	/* CPU frequency resume boost in kHz */
-	unsigned int boost_resume_reason;
+	unsigned long long boost_resume_reason;
 };
 
 /* Tegra io dpd entry - for each supported driver */
